update leet: [ 526, 576 ]

+++ OJ/leet526/leet526-状态压缩DP.cpp
@@ -16,12 +16,12 @@ public:
         // f[010101] 代表数字 1 3 5 放在前三位且合法的排列的数目
         // 转移: f[010111] += f[010101] 指的是 2 放入第 4 位（合法）
         vector<int> f(1 << n);
-        for (size_t i = 0; i < n; i++)
+        for (int i = 0; i < n; i++)
         {
             f[1 << i] = 1;
         }
 
-        for (size_t i = 0; i < (1 << n); i++)
+        for (int i = 0; i < (1 << n); i++)
         {
             // 计算 i 中 1 的数目
             int cnt = 1, x = i;
@@ -29,7 +29,7 @@ public:
                 x &= x-1; ++cnt;
             }
 
-            for (size_t j = 0; j < n; j++)
+            for (int j = 0; j < n; j++)
             {
                 if (not(i & (1<<j)) and (
                     (j + 1) % cnt == 0 || cnt % (j + 1) == 0
